Removed old files and moved over rest of server files to typescript
